Re-add the managed client to the build task.
Brings back the managed client and reworks the two new build
task specific tests to compile using the managed client inside
the build task.

+++ src/Compilers/Core/VBCSCompiler/CompilerRequestHandler.cs
@@ -94,10 +94,10 @@ namespace Microsoft.CodeAnalysis.CompilerServer
                 }
                 else if (arg.ArgumentId == BuildProtocolConstants.ArgumentId.CommandLineArgument)
                 {
-                    uint argIndex = arg.ArgumentIndex;
+                    int argIndex = arg.ArgumentIndex;
                     while (argIndex >= commandLineArguments.Count)
                         commandLineArguments.Add("");
-                    commandLineArguments[(int)argIndex] = arg.Value;
+                    commandLineArguments[argIndex] = arg.Value;
                 }
             }
